AutoPark/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt dev-requirements.in requirements.in
#
asgiref==3.7.2
    # via django
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
django==5.0.1
    # via
    #   -r requirements.in
    #   django-bootstrap5
    #   djangorestframework
django-bootstrap5==23.4
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-gis
djangorestframework-gis==1.0
    # via -r requirements.in
exceptiongroup==1.2.0
    # via pytest
factory-boy==3.3.0
    # via -r requirements.in
faker==24.1.0
    # via
    #   factory-boy
    #   mixer
geographiclib==2.0
    # via geopy
geopy==2.4.1
    # via -r requirements.in
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
mixer==7.2.0
    # via -r dev-requirements.in
openrouteservice==2.3.3
    # via -r requirements.in
packaging==24.0
    # via pytest
pluggy==1.4.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pytest==8.1.1
    # via
    #   -r dev-requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r dev-requirements.in
python-dateutil==2.9.0.post0
    # via faker
pytz==2023.3.post1
    # via djangorestframework
requests==2.31.0
    # via openrouteservice
ruff==0.3.7
    # via -r dev-requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via pytest
typing-extensions==4.9.0
    # via asgiref
urllib3==2.2.1
    # via requests
